Moved SyncRpc initialization to node module
