+++ packages/lodestar/test/unit/sync/initial.test.ts
@@ -5,7 +5,7 @@ import {Libp2pNetwork} from "../../../src/network";
 import {BeaconDB} from "../../../src/db/api";
 import {ReputationStore} from "../../../src/sync/reputation";
 import {WinstonLogger} from "../../../src/logger";
-import {SyncRpc} from "../../../src/sync/rpc";
+import {SyncRpc} from "../../../src/network/libp2p/syncRpc";
 import {InitialSync} from "../../../src/sync/initial";
 import {generateState} from "../../utils/state";
 import PeerInfo from "peer-info";
